+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Fixed crash on negative aspect count
+- Slightly improve the aspect cache creation
 
 ## 1.1.1
 - Fixed crash on dedicated servers